Implement URI API for run-time configuration (#13)
Supported URI query parameters by sqleet:

  * salt: Provides 16-byte salt for the key derivation function (KDF)
  * header: Overrides 16-bytes in the beginning of the database header
  * kdf: Specifies KDF algorithm (only `none` value supported for now)
  * skip: Run-time setting overriding compile-time SKIP_HEADER_BYTES
  * page_size: Equivalent to `page_size` PRAGMA setting

Parameters `salt` and `header` have corresponding hex-prefixed versions.
Both expect 16-byte values and shorter strings are 0-padded to 16 bytes.
The KDF salt is stored in the first 16 bytes of the database header if
`header` is undefined. Otherwise the value of `header` overwrites the 16
header bytes (the KDF salt can thus be hidden from the database file).

Old raw key interface is deprecated by `kdf=none` that disables the KDF.
When KDF is disabled, `key` and `rekey` PRAGMAs (and corresponding URI
parameters) accept a 32-byte value that becomes the "master" key which
is normally derived by the KDF. This allows the users of the library to
take full control of the key derivation process if needed.

There is not yet a user-friendly method to retrieve the currently used
salt or other settings. Moreover, changing the settings of an existing
database after its creation remains cumbersome. It can however be done
by first fully decrypting the database (with `PRAGMA rekey=''`) and then
encrypting it again (with `PRAGMA rekey='password'`) after re-opening
the database file with new URI parameters. Addressing these shortcomings
is a low-priority long-term goal given the workarounds and the lack of
low-complexity solutions to them without major code refactoring.

README.md documentation of the URI API is left for a later commit.