Implement URI API for run-time configuration (#13)
Supported URI query parameters by sqleet:

  * salt: Provides 16-byte salt for the key derivation function (KDF)
  * header: Overrides 16-bytes in the beginning of the database header
  * kdf: Specifies KDF algorithm (only `none` value supported for now)
  * skip: Run-time setting overriding compile-time SKIP_HEADER_BYTES
  * page_size: Equivalent to `page_size` PRAGMA setting

Parameters `salt` and `header` have corresponding hex-prefixed versions.
Both expect 16-byte values and shorter strings are 0-padded to 16 bytes.
The KDF salt is stored in the first 16 bytes of the database header if
`header` is undefined. Otherwise the value of `header` overwrites the 16
header bytes (the KDF salt can thus be hidden from the database file).

Old raw key interface is deprecated by `kdf=none` that disables the KDF.
When KDF is disabled, `key` and `rekey` PRAGMAs (and corresponding URI
parameters) accept a 32-byte value that becomes the "master" key which
is normally derived by the KDF. This allows the user of the library to
take full control of the key derivation process.

`VACUUM INTO` (introduced in SQLite 3.27.0) supports changing the URI
settings of an existing database by overriding the URI parameter values
in the `INTO` URI filename. Unspecified parameters inherit their values
from the main database vacuumed by the command.

There is not yet a user-friendly method to retrieve the currently used
salt or other settings as required by #13. This is a low-priority task.