Fix page buffer memory leak if setting codec twice
A page buffer of a codec is allocated when the codec is set to Btree.
However, if the same codec was set to Btree again, then the pointer to
the page buffer got simply overwitten with a pointer to newly allocated
page buffer. Thus, the old page buffer memory was leaked. This commit
adds missing sqlite3_free() call when setting a codec multiple times.